Use PHOTON_ENABLE_FUSE to select between libfuse2 and libfuse3
