+++ CMake/Findfuse.cmake
@@ -1,18 +1,7 @@
-find_library(FUSE3_LIBRARIES fuse3)
-find_path(FUSE3_INCLUDE_DIRS fuse3/fuse.h)
-if(FUSE3_LIBRARIES AND FUSE3_INCLUDE_DIRS)
-    message(STATUS "Found libfuse3 at ${FUSE3_INCLUDE_DIRS}")
-    set(FUSE_LIBRARIES ${FUSE3_LIBRARIES})
-    set(FUSE_INCLUDE_DIRS ${FUSE3_INCLUDE_DIRS})
-    set(FUSE_COMPILE_FLAGS _FILE_OFFSET_BITS=64 FUSE_USE_VERSION=35)
-else()
-    message(STATUS "libfuse3 not found, trying to find libfuse2...")
-    find_library(FUSE_LIBRARIES fuse)
-    find_path(FUSE_INCLUDE_DIRS fuse.h)
-    if(FUSE_LIBRARIES AND FUSE_INCLUDE_DIRS)
-        message(STATUS "Found libfuse2 at ${FUSE_INCLUDE_DIRS}")
-        set(FUSE_COMPILE_FLAGS _FILE_OFFSET_BITS=64 FUSE_USE_VERSION=29)
-    else()
-        message(FATAL_ERROR "Neither libfuse3 nor libfuse2 found")
-    endif()
-endif()
+find_path(FUSE_INCLUDE_DIRS fuse.h)
+
+find_library(FUSE_LIBRARIES fuse)
+
+find_package_handle_standard_args(fuse DEFAULT_MSG FUSE_LIBRARIES FUSE_INCLUDE_DIRS)
+
+mark_as_advanced(FUSE_INCLUDE_DIRS FUSE_LIBRARIES)
