+++ CMakeLists.txt
@@ -19,7 +19,7 @@ set(PHOTON_CXX_STANDARD "14" CACHE STRING "C++ standard")
 option(PHOTON_BUILD_TESTING "enable build testing" OFF)
 option(PHOTON_BUILD_WITH_ASAN "build with asan" OFF)
 option(PHOTON_ENABLE_URING "enable io_uring function" OFF)
-option(PHOTON_ENABLE_FUSE "enable fuse function" OFF)
+set(PHOTON_ENABLE_FUSE "OFF" CACHE STRING "enable fuse version, OFF for disable, ON/2 for libfuse2, 3 for libfuse3")
 option(PHOTON_GLOBAL_INIT_OPENSSL "Turn this off if any of your third-party libs inits old-version OpenSSL as well,
 because Photon will register coroutine locks for crypto. But don't bother if you have latest OpenSSL >= 1.1.0" ON)
 option(PHOTON_ENABLE_SASL "enable sasl" OFF)
@@ -129,8 +129,10 @@ set(dependencies zlib openssl)
 
 if (CMAKE_SYSTEM_NAME MATCHES "Linux")
     LIST(APPEND dependencies aio)
-    if (PHOTON_ENABLE_FUSE)
+    if (PHOTON_ENABLE_FUSE STREQUAL "ON" OR PHOTON_ENABLE_FUSE STREQUAL "2")
         LIST(APPEND dependencies fuse)
+    elseif (PHOTON_ENABLE_FUSE STREQUAL "3")
+        LIST(APPEND dependencies fuse3)
     endif ()
     if (PHOTON_ENABLE_URING)
         LIST(APPEND dependencies uring)
@@ -212,8 +214,8 @@ else ()
         list(APPEND PHOTON_SRC io/iouring-wrapper.cpp)
     endif ()
 endif ()
-if (PHOTON_ENABLE_FUSE)
-    list(APPEND PHOTON_SRC io/fuse-adaptor.cpp fs/fuse_adaptor.cpp)
+if (PHOTON_ENABLE_FUSE STREQUAL "ON" OR PHOTON_ENABLE_FUSE STREQUAL "2" OR PHOTON_ENABLE_FUSE STREQUAL "3")
+    list(APPEND PHOTON_SRC fs/fuse_adaptor.cpp)
 endif ()
 if (PHOTON_ENABLE_SASL)
     list(APPEND PHOTON_SRC net/security-context/sasl-stream.cpp)
@@ -246,9 +248,12 @@ target_include_directories(photon_obj PRIVATE include ${OPENSSL_INCLUDE_DIRS} ${
         ${ZLIB_INCLUDE_DIRS} ${CURL_INCLUDE_DIRS}
 )
 
-if (PHOTON_ENABLE_FUSE)
-    target_compile_definitions(photon_obj PRIVATE ${FUSE_COMPILE_FLAGS})
+if (PHOTON_ENABLE_FUSE STREQUAL "ON" OR PHOTON_ENABLE_FUSE STREQUAL "2")
+    target_compile_definitions(photon_obj PRIVATE _FILE_OFFSET_BITS=64 FUSE_USE_VERSION=29)
     target_include_directories(photon_obj PRIVATE ${FUSE_INCLUDE_DIRS})
+elseif (PHOTON_ENABLE_FUSE STREQUAL "3")
+    target_compile_definitions(photon_obj PRIVATE _FILE_OFFSET_BITS=64 FUSE_USE_VERSION=35)
+    target_include_directories(photon_obj PRIVATE ${FUSE3_INCLUDE_DIRS})
 endif ()
 if (PHOTON_GLOBAL_INIT_OPENSSL)
     target_compile_definitions(photon_obj PRIVATE PHOTON_GLOBAL_INIT_OPENSSL)
@@ -299,8 +304,10 @@ endif ()
 if (PHOTON_ENABLE_URING)
     list(APPEND static_deps ${URING_LIBRARIES})
 endif ()
-if (PHOTON_ENABLE_FUSE)
+if (PHOTON_ENABLE_FUSE STREQUAL "ON" OR PHOTON_ENABLE_FUSE STREQUAL "2")
     list(APPEND static_deps ${FUSE_LIBRARIES})
+elseif (PHOTON_ENABLE_FUSE STREQUAL "3")
+    list(APPEND static_deps ${FUSE3_LIBRARIES})
 endif ()
 if (PHOTON_ENABLE_SASL)
     list(APPEND static_deps ${GSASL_LIBRARIES})
