+++ fs/fuse_adaptor.h
@@ -42,5 +42,7 @@ void set_fuse_fs(IFileSystem* fs);
 
 fuse_operations* get_fuse_xmp_oper();
 
+int run_fuse(int argc, char *argv[], const struct fuse_operations *op,
+    void *user_data);
 }
 }
