+++ include/photon/io/fuse-adaptor.h
@@ -1,1 +1,1 @@
-../../../io/fuse-adaptor.h
+../../../fs/fuse_adaptor.h
